# api/docs/swagger.yaml
definitions:
...
  http.Response:
    properties:
      data: {}
      description:
        type: string
      error: {}
      message:
        type: string
      requestId:
        type: string
      status:
        type: string
    type: object
...
  models.Entity:
    properties:
      _id:
        type: string
      created_at:
        type: string
      description:
        type: string
      name:
        type: string
      slug:
        type: string
      status:
        type: integer
      updated_at:
        type: string
    type: object
...
  models.JoinGroupRequest:
    properties:
      _id:
        type: string
      from_date:
        type: string
      lookups:
        items:
          $ref: '#/definitions/models.Lookup'
        type: array
      name:
        type: string
      to_date:
        type: string
    type: object
  models.Lookup:
    properties:
      as:
        type: string
      foreignField:
        type: string
      from:
        type: string
      localField:
        type: string
    type: object
...
info:
  contact: {}
  description: This is a api gateway
  termsOfService: https://kassa360.uz
paths:
...
  /client-api/entity/{slug}/create:
    post:
      consumes:
      - application/json
      description: Create Entity
      operationId: create_entity
      parameters:
      - default: '"Asia/Tashkent"'
        description: Location
        in: header
        name: location
        type: string
      - description: Entity Slug
        in: path
        name: slug
        required: true
        type: string
      - description: Create Entity
        in: body
        name: entity
        required: true
        schema:
          $ref: '#/definitions/models.Entity'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Entity'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: Create Entity
      tags:
      - Entity
  /client-api/entity/{slug}/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Entity
      operationId: delete_entity
      parameters:
      - default: '"Asia/Tashkent"'
        description: Location
        in: header
        name: location
        type: string
      - description: Entity Slug
        in: path
        name: slug
        required: true
        type: string
      - description: Entity ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Deleted
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: Delete Entity
      tags:
      - Entity
  /client-api/entity/{slug}/get-all:
    post:
      consumes:
      - application/json
      description: Get All Entities
      operationId: get_all
      parameters:
      - default: '"Asia/Tashkent"'
        description: Location
        in: header
        name: location
        type: string
      - description: Entity Slug
        in: path
        name: slug
        required: true
        type: string
      - description: Offset
        in: query
        name: offset
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Sort
        in: query
        name: sort
        type: string
      - description: Order
        in: query
        name: order
        type: string
      - description: Search
        in: query
        name: search
        type: string
      - description: Filter Entity
        in: body
        name: filter
        schema:
          $ref: '#/definitions/models.Entity'
      produces:
      - application/json
      responses:
        "200":
          description: Entities
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Entity'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get All Entities
      tags:
      - Entity
  /client-api/entity/{slug}/get-join:
    post:
      consumes:
      - application/json
      description: Get Join Entities
      operationId: get_join
      parameters:
      - default: '"Asia/Tashkent"'
        description: Location
        in: header
        name: location
        type: string
      - description: Entity Slug
        in: path
        name: slug
        required: true
        type: string
      - description: Offset
        in: query
        name: offset
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Sort
        in: query
        name: sort
        type: string
      - description: Order
        in: query
        name: order
        type: string
      - description: Search
        in: query
        name: search
        type: string
      - description: Filter Entity
        in: body
        name: filter
        schema:
          $ref: '#/definitions/models.JoinGroupRequest'
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Get Join Entities
      tags:
      - Entity
  /client-api/entity/{slug}/get/{id}:
    get:
      consumes:
      - application/json
      description: Get Entity
      operationId: get_entity
      parameters:
      - default: '"Asia/Tashkent"'
        description: Location
        in: header
        name: location
        type: string
      - description: Entity Slug
        in: path
        name: slug
        required: true
        type: string
      - description: Entity ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Entity'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: Get Entity
      tags:
      - Entity
  /client-api/entity/{slug}/update/{id}:
    put:
      consumes:
      - application/json
      description: Update Entity
      operationId: update_entity
      parameters:
      - description: Entity Slug
        in: path
        name: slug
        required: true
        type: string
      - description: Entity ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Entity
        in: body
        name: entity
        required: true
        schema:
          $ref: '#/definitions/models.Entity'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Entity'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                data:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: Update Entity
      tags:
      - Entity
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  BasicAuth:
    type: basic
swagger: "2.0"